requirements.txt
yt-dlp>=2023.12.30
openai-whisper>=20231117
transformers>=4.36.0
peft>=0.7.0
datasets>=2.14.0
torch>=2.1.0
torchaudio>=2.1.0
accelerate>=0.24.0
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.1
sacrebleu>=2.3.1
